handle docker images cleaning

+++ src/deploy/osp_deployer/director.py
@@ -524,7 +524,7 @@ class Director(InfraHost):
         domain_param = "  CloudDomain:"
         rbd_backend_param = "  NovaEnableRbdBackend:"
         glance_backend_param = "  GlanceBackend:"
-        rbd_cinder_backend_param = "  CinderEnableRbdBackend:" 
+        rbd_cinder_backend_param = "  CinderEnableRbdBackend:"
         osds_per_node = 0
 
         if osd_disks:
@@ -710,7 +710,8 @@ class Director(InfraHost):
     def setup_dell_storage(self):
 
         # Clean the local docker registry
-        self.run_tty("sudo docker rmi $(sudo docker images -a -q) --force")
+        if len(self.run_tty("sudo docker images -a -q")[0]) > 1:
+            self.run_tty("sudo docker rmi $(sudo docker images -a -q) --force")
 
         # Re - Upload the yaml files in case we're trying to
         # leave the undercloud intact but want to redeploy with
@@ -873,7 +874,7 @@ class Director(InfraHost):
             'docker pull ' + remote_url,
             'docker tag ' + remote_url + ' ' + local_url,
             'docker push ' + local_url,
-            'sed -i "50i \  DockerManilaShareImage: ' + local_url +	
+            'sed -i "50i \  DockerManilaShareImage: ' + local_url +
             '" ' + overcloud_images_file,
             'echo "  DockerInsecureRegistryAddress:" >> ' +
             overcloud_images_file,
